% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetALLHazLogSlope.R
\name{GetALLHazLogSlope}
\alias{GetALLHazLogSlope}
\title{Computes the posterior distribution of hazard value for a vector x for the Piecewise Linear Log Hazard model (PLLH)}
\usage{
GetALLHazLogSlope(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the posterior mean hazard function}

\item{G1}{List of posterior samples from the BayesPiecewiseLinearLogHazard function.}
}
\value{
Matrix containing the posterior distribution of hazard values h(x)
}
\description{
Computes the posterior distribution of hazard value for a vector x for the Piecewise Linear Log Hazard model (PLLH)
}
