% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetALLHazPiece.R
\name{GetALLHazPiece}
\alias{GetALLHazPiece}
\title{Computes the posterior hazard values for a vector x for the Piecewise Exponential Hazard model (PEH)}
\usage{
GetALLHazPiece(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the  hazard.}

\item{G1}{List of posterior samples from the BayesPiecewiseHazard function.}
}
\value{
Matrix containing the posterior distribution of hazard values h(x)
}
\description{
Computes the posterior hazard values for a vector x for the Piecewise Exponential Hazard model (PEH)
}
