% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetALLSurvPEH.R
\name{GetALLSurvPEH}
\alias{GetALLSurvPEH}
\title{Computes the posterior distribution of survival probabilities for a vector x for the Piecewise Exponential Hazard model (PEH)}
\usage{
GetALLSurvPEH(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the posterior mean survival probability.}

\item{G1}{List of posterior samples from the BayesPiecewiseLinearHazard function.}
}
\value{
Matrix containing the posterior distribution of survival probabilities S(x)
}
\description{
Computes the posterior distribution of survival probabilities for a vector x for the Piecewise Exponential Hazard model (PEH)
}
