% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetALLSurvPLLH.R
\name{GetALLSurvPLLH}
\alias{GetALLSurvPLLH}
\title{Computes posterior distribution of survival probabilities for a vector x for the Piecewise Linear Log Hazard model (PLLH)}
\usage{
GetALLSurvPLLH(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the posterior mean survival probability.}

\item{G1}{List of posterior samples from the BayesPiecewiseLinearLogHazard function.}
}
\value{
Matrix containing the posterior distribution survival probabilities S(x)
}
\description{
Computes posterior distribution of survival probabilities for a vector x for the Piecewise Linear Log Hazard model (PLLH)
}
