% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostMeanHazLogSlope.R
\name{PostMeanHazLogSlope}
\alias{PostMeanHazLogSlope}
\title{Computes the posterior mean hazard value for a vector x for the Piecewise Linear Log Hazard model (PLLH)}
\usage{
PostMeanHazLogSlope(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the posterior mean hazard function}

\item{G1}{List of posterior samples from the BayesPiecewiseLinearLogHazard function.}
}
\value{
Vector containing the posterior mean hazard values h(x)
}
\description{
Computes the posterior mean hazard value for a vector x for the Piecewise Linear Log Hazard model (PLLH)
}
