% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostMeanHazPiece.R
\name{PostMeanHazPiece}
\alias{PostMeanHazPiece}
\title{Computes the posterior mean hazard values for a vector x for the Piecewise Exponential Hazard model (PEH)}
\usage{
PostMeanHazPiece(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the posterior mean hazard.}

\item{G1}{List of posterior samples from the BayesPiecewiseHazard function.}
}
\value{
Vector containing the posterior mean hazard values h(x)
}
\description{
Computes the posterior mean hazard values for a vector x for the Piecewise Exponential Hazard model (PEH)
}
