% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostMeanSurvPLLH.R
\name{PostMeanSurvPLLH}
\alias{PostMeanSurvPLLH}
\title{Computes the posterior mean survival probabilities for a vector x for the Piecewise Linear Log Hazard model (PLLH)}
\usage{
PostMeanSurvPLLH(x, G1)
}
\arguments{
\item{x}{Vector of times to compute the posterior mean survival probability.}

\item{G1}{List of posterior samples from the BayesPiecewiseLinearLogHazard function.}
}
\value{
Vector containing the posterior mean survival probabilities S(x)
}
\description{
Computes the posterior mean survival probabilities for a vector x for the Piecewise Linear Log Hazard model (PLLH)
}
